Added vector doc

+++ src/com/blackoutburst/bogel/maths/Vector3i.java
@@ -1,61 +1,178 @@
 package com.blackoutburst.bogel.maths;
 
+/**
+ * <h1>Vector3i</h1>
+ * 
+ * <p>
+ * Create and manager Vector3
+ * </p>
+ * 
+ * @since 0.1
+ * @author Blackoutburst
+ */
 public class Vector3i {
 
+	/**x value*/
 	public int x;
+	
+	/**y value*/
 	public int y;
+	
+	/**z value*/
 	public int z;
 	
+	/**
+	 * <p>
+	 * Create a new vector (0, 0, 0)
+	 * </p>
+	 * 
+	 * @since 0.1
+	 * @author Blackoutburst
+	 */
 	public Vector3i() {
 		this.x = 0;
 		this.y = 0;
 		this.z = 0;
 	}
 	
+	/**
+	 * <p>
+	 * Create a new vector (size, size, size)
+	 * </p>
+	 *
+	 * @param int size
+	 * @since 0.1
+	 * @author Blackoutburst
+	 */
 	public Vector3i(int size) {
 		this.x = size;
 		this.y = size;
 		this.z = size;
 	}
 	
+	/**
+	 * <p>
+	 * Create a new vector (x, y, z)
+	 * </p>
+	 *
+	 * @param int x
+	 * @param int y
+	 * @param int z
+	 * @since 0.1
+	 * @author Blackoutburst
+	 */
 	public Vector3i(int x, int y, int z) {
 		this.x = x;
 		this.y = y;
 		this.z = z;
 	}
 	
+	/**
+	 * <p>
+	 * Set vector values (x, y, z)
+	 * </p>
+	 *
+	 * @param int x
+	 * @param int y
+	 * @param int z
+	 * @since 0.1
+	 * @author Blackoutburst
+	 */
 	public void set(int x, int y, int z) {
 		this.x = x;
 		this.y = y;
 		this.z = z;
 	}
 	
+	/**
+	 * <p>
+	 * Get vector X value
+	 * </p>
+	 *
+	 * @return int
+	 * @since 0.1
+	 * @author Blackoutburst
+	 */
 	public int getX() {
 		return (x);
 	}
 
+	/**
+	 * <p>
+	 * Set vector X value
+	 * </p>
+	 *
+	 * @param int x
+	 * @since 0.1
+	 * @author Blackoutburst
+	 */
 	public void setX(int x) {
 		this.x = x;
 	}
 
+	/**
+	 * <p>
+	 * Get vector Y value
+	 * </p>
+	 *
+	 * @return int
+	 * @since 0.1
+	 * @author Blackoutburst
+	 */
 	public int getY() {
 		return (y);
 	}
 
+	/**
+	 * <p>
+	 * Set vector Y value
+	 * </p>
+	 *
+	 * @param int y
+	 * @since 0.1
+	 * @author Blackoutburst
+	 */
 	public void setY(int y) {
 		this.y = y;
 	}
 	
+	/**
+	 * <p>
+	 * Get vector Z value
+	 * </p>
+	 *
+	 * @return int
+	 * @since 0.1
+	 * @author Blackoutburst
+	 */
 	public int getZ() {
 		return (z);
 	}
 
+	/**
+	 * <p>
+	 * Set vector Z value
+	 * </p>
+	 *
+	 * @param int z
+	 * @since 0.1
+	 * @author Blackoutburst
+	 */
 	public void setZ(int z) {
 		this.z = z;
 	}
 	
+	/**
+	 * <p>
+	 * Normalize the vector
+	 * </p>
+	 *
+	 * @return Vector3i
+	 * @since 0.1
+	 * @author Blackoutburst
+	 */
 	public Vector3i normalize() {
-		int mag = (int) Math.sqrt(x * x + y * y + z * z);
+		float mag = (float) Math.sqrt(x * x + y * y + z * z);
 		x /= mag;
 		y /= mag;
 		z /= mag;
@@ -63,7 +180,18 @@ public class Vector3i {
 		return (this);
 	}
 	
-	public Vector3i mul(int v) {
+	
+	/**
+	 * <p>
+	 * Multiply two vector
+	 * </p>
+	 *
+	 * @param float v
+	 * @return Vector3i
+	 * @since 0.1
+	 * @author Blackoutburst
+	 */
+	public Vector3i mul(float v) {
 		x *= v;
 		y *= v;
 		z *= v;
@@ -71,7 +199,16 @@ public class Vector3i {
 		return (this);
 	}
 	
-	public int length() {
-		return ((int) Math.sqrt(x * x + y * y + z * z));
+	/**
+	 * <p>
+	 * 	Return the vector length
+	 * </p>
+	 *
+	 * @return float
+	 * @since 0.1
+	 * @author Blackoutburst
+	 */
+	public float length() {
+		return ((float) Math.sqrt(x * x + y * y + z * z));
 	}
 }
